+++ src/main/java/utils/Statistics.java
@@ -6,10 +6,10 @@ public class Statistics {
 
     //private static String[]encabezado = {"General", "Light Aircraft", "Mid Weight Aircraft", "Heavy Aircraft", "Maintenance"};
     private static int[]idCount = new int[classesNumber];
-    private static int[]totalWaitingTime = new int[classesNumber];
-    private static int[]maxWaitingTime = new int[classesNumber];
-    private static int[]totalTransitTime = new int[classesNumber];
-    private static int[]maxTransitTime = new int[classesNumber];
+    private static double[]totalWaitingTime = new double[classesNumber];
+    private static double[]maxWaitingTime = new double[classesNumber];
+    private static double[]totalTransitTime = new double[classesNumber];
+    private static double[]maxTransitTime = new double[classesNumber];
 
     public static int getIdCount(int classEntityId) {
         return idCount[classEntityId];
@@ -19,35 +19,35 @@ public class Statistics {
         Statistics.idCount[classEntityId]++;
     }
 
-    public static int getTotalWaitingTime(int classEntityId) {
+    public static double getTotalWaitingTime(int classEntityId) {
         return totalWaitingTime[classEntityId];
     }
 
-    public static void accumulateWaitingTime(int WaitingTime, int classEntityId) {
+    public static void accumulateWaitingTime(double WaitingTime, int classEntityId) {
         Statistics.totalWaitingTime[classEntityId] += WaitingTime;
     }
 
-    public static int getMaxWaitingTime(int classEntityId) {
+    public static double getMaxWaitingTime(int classEntityId) {
         return maxWaitingTime[classEntityId];
     }
 
-    public static void setMaxWaitingTime(int maxWaitingTime, int classEntityId) {
+    public static void setMaxWaitingTime(double maxWaitingTime, int classEntityId) {
         Statistics.maxWaitingTime[classEntityId] = maxWaitingTime;
     }
 
-    public static int getTotalTransitTime(int classEntityId) {
+    public static double getTotalTransitTime(int classEntityId) {
         return totalTransitTime[classEntityId];
     }
 
-    public static void accumulateTransitTime(int TransitTime, int classEntityId) {
+    public static void accumulateTransitTime(double TransitTime, int classEntityId) {
         Statistics.totalTransitTime[classEntityId] += TransitTime;
     }
 
-    public static int getMaxTransitTime(int classEntityId) {
+    public static double getMaxTransitTime(int classEntityId) {
         return maxTransitTime[classEntityId];
     }
 
-    public static void setMaxTransitTime(int maxTransitTime, int classEntityId) {
+    public static void setMaxTransitTime(double maxTransitTime, int classEntityId) {
         Statistics.maxTransitTime[classEntityId] = maxTransitTime;
     }
 }
